- multispecies with ACECTildeEvaluator - modify CMaleLists.txt - add yaml-cpp into lib/user-pace/yaml-cpp

+++ src/USER-PACE/ace_types.h
@@ -39,7 +39,7 @@ typedef short int DENSITY_TYPE;
 
 typedef short int MS_TYPE;
 
-typedef short int SHORT_INT_TYPE;
+typedef int SHORT_INT_TYPE;
 typedef double DOUBLE_TYPE;
 
 #endif
